+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+## V1.1.0 - 2020-1-12
+
+### Added
+
+- `"createVirtualRoomBlind"` Creates a virtual blind for each room that contains all the blinds in the room. Default is true. Changing this blind will change all blinds in the room.
+
+- `"createDiscreteBlinds"` Creates a blind for each real blind. Default is true. You might want to turn this off if you want only the virtual room blind.
+
+- `"prefixRoomNameToBlindName"` Prefix the room name to each discrete blind name. Default is true.
+
 ## v1.0.0 - 2020-1-4
 
 ### Added